# swagger/swagger-fuhrparkmanagement.yaml
openapi: "3.0.3"
info:
  title: "Fuhrparkmanagement API"
  description: "Fuhrparkmanagement API"
  version: "1.0.0"
servers:
  - url: "http://localhost:8080"
...
paths:
  /cars:
    get:
      tags:
        - Cars
      summary: "GET cars"
      operationId: "listAll"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Car"
        "403":
          description: "Forbidden"
    post:
      tags:
        - Cars
      summary: "POST cars"
      operationId: "create"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Car"
        required: true
      responses:
        "201":
          description: "Created"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/Car"
        "400":
          description: "Bad Request"
        "403":
          description: "Forbidden"
...
  /rides:
    get:
      tags:
        - Rides
      summary: "GET rides"
      operationId: "listAll"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ride"
        "403":
          description: "Forbidden"
    post:
      tags:
        - Rides
      summary: "POST rides"
      operationId: "create"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Ride"
        required: true
      responses:
        "201":
          description: "Created"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/Ride"
        "400":
          description: "Bad Request"
        "403":
          description: "Forbidden"
...
  /reservations:
    get:
      tags:
        - Reservations
      summary: "GET reservations"
      operationId: "listAll"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
        "403":
          description: "Forbidden"
    post:
      tags:
        - Reservations
      summary: "POST reservations"
      operationId: "create"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Reservation"
        required: true
      responses:
        "201":
          description: "Created"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/Reservation"
        "400":
          description: "Bad Request"
        "403":
          description: "Forbidden"
...
  /statistics/users:
    get:
      tags:
        - Statistics
      summary: "GET statistics/users"
      operationId: "listAll"
      parameters:
        - name: "currentMonth"
          in: "query"
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserStatistic"
        "403":
          description: "Forbidden"
...
components:
  schemas:
    Car:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          readOnly: true
        licensePlate:
          type: "string"
        modelName:
          type: "string"
        fuelType:
          type: "string"
          enum:
            - "PETROL"
            - "DIESEL"
            - "ELECTRIC"
        location:
          type: "string"
        seats:
          type: "integer"
          format: "int32"
          minimum: 1
        range:
          type: "integer"
          format: "int32"
          minimum: 1
        available:
          type: "boolean"
          default: true
        createdDate:
          type: string
          format: date-time
          readOnly: true
        deletedDate:
          type: string
          format: date-time
          readOnly: true
      required:
        - licensePlate
        - modelName
        - fuelType
        - location
        - seats
        - range
    Ride:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          readOnly: true
        carId:
          type: "integer"
          format: "int64"
        startAddress:
          type: "string"
        destinationAddress:
          type: "string"
        startDate:
          type: "string"
          format: "date-time"
        endDate:
          type: "string"
          format: "date-time"
        purpose:
          type: "string"
        createdDate:
          type: string
          format: date-time
          readOnly: true
        deletedDate:
          type: string
          format: date-time
          readOnly: true
      required:
        - carId
        - startAddress
        - destinationAddress
        - startDate
        - endDate
        - purpose
    Reservation:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          readOnly: true
        userId:
          type: "string"
        rideId:
          type: "integer"
          format: "int64"
        createdDate:
          type: string
          format: date-time
          readOnly: true
        deletedDate:
          type: string
          format: date-time
          readOnly: true
      required:
        - userId
        - rideId
    UserStatistic:
      type: object
      properties:
        userId:
          type: string
        name:
          type: string
        amountReservations:
          type: integer
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []